just finished ir56e_pc_e, ir56e_pc_c

+++ app/Helpers/Forms/CommencementFormPdfHelper.php
@@ -5,10 +5,16 @@ use setasign\Fpdi\Fpdi;
 
 use App\Helpers\FormPdf;
 
+use App\Models\IrdFormFileField;
+
 class CommencementFormPdfHelper {
   public static $yOffset = 2;
 
-  public static function generate($data, $templateFilePath, $finalFilePath, $fields=[], $mappings=[] ) {
+  public static function generate($data, $finalFilePath, $irdFormFile, $mappings=[] ) {
+    $templateFilePath = storage_path('forms/'.$irdFormFile->file);
+
+    self::$yOffset = $irdFormFile->y_offset;
+
     $pdf = new FormPdf();
     $pdf->AddPage();
     $pdf->setSourceFile( $templateFilePath );
@@ -27,67 +33,177 @@ class CommencementFormPdfHelper {
     //Cell($w, $h=0, $txt='', $border=0, $ln=0, $align='', $fill=0, $link='', $stretch=0, $ignore_min_height=false, $calign='T', $valign='M')
 
 
-    $fields = [
+//    $irdForm = $irdFormFile->irdForm;
+//    $irdFormFileEng = $irdForm->files()->whereLangId(3)->first();
+//    $irdFormFileFieldsEng = $irdFormFileEng->fields;
+//
+//    foreach($irdFormFileFieldsEng as $item) {
+//      $irdFormFile->fields()->save(new IrdFormFileField([
+//        'key'=>$item->key,
+//        'type'=>$item->type,
+//        'x'=>$item->x,
+//        'y'=>$item->y,
+//        'font_size'=>$item->font_size,
+//        'relative_to'=>$item->relative_to,
+//        'relative_to_key_id'=>$item->relative_to_key_id,
+//        'width'=>$item->width,
+//        'field_count'=>$item->field_count,
+//        'align'=>$item->align,
+//        'char_align'=>$item->char_align,
+//        'lang'=>$item->lang,
+//        'append_asterisk'=>$item->append_asterisk,
+//        'remark'=>''
+//      ]));
+//    }
+//    dd('ok');
+    $fieldsxxx = [
       // company
-      ['key'=>'fileNo','type'=>'string', 'x'=>52, 'y'=>44, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'ern','type'=>'string', 'x'=>52, 'y'=>48, 'fontSize'=>10, 'width'=>100, 'fontStyle'=>'B'],
-      ['key'=>'erName','type'=>'string', 'x'=>52, 'y'=>52, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'erAddress','type'=>'string', 'x'=>52, 'y'=>56, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'signatureName','type'=>'string', 'x'=>52, 'y'=>60, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'designation','type'=>'string', 'x'=>52, 'y'=>64, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'formDate','type'=>'string', 'x'=>52, 'y'=>68, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'fileNo','type'=>'string', 'x'=>65, 'y'=>37, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'ern','type'=>'string', 'x'=>100, 'y'=>37, 'fontSize'=>12, 'width'=>90, 'appendAsterisk'=>true],
+      ['key'=>'erName','type'=>'string', 'x'=>65, 'y'=>41.2, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'erAddress','type'=>'string', 'x'=>65, 'y'=>45.4, 'fontSize'=>10, 'width'=>100],
+
       // employee
-      ['key'=>'name','type'=>'string', 'x'=>52, 'y'=>72, 'fontSize'=>10, 'width'=>100, 'fontStyle'=>'B'],
-      ['key'=>'nameInChinese','type'=>'string', 'x'=>52, 'y'=>80, 'fontSize'=>10, 'width'=>100, 'lang'=>'chn'],
-      ['key'=>'hkid','type'=>'string', 'x'=>52, 'y'=>84, 'fontSize'=>10, 'width'=>100, 'fontStyle'=>'B'],
-      ['key'=>'ppNum','type'=>'string', 'x'=>52, 'y'=>88, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'gender','type'=>'string', 'x'=>52, 'y'=>92, 'fontSize'=>10, 'width'=>100, 'fontStyle'=>'B'],
-      ['key'=>'martialStatus','type'=>'string', 'x'=>52, 'y'=>96, 'fontSize'=>10, 'width'=>100, 'fontStyle'=>'B'],
+      ['key'=>'name','type'=>'string', 'x'=>100, 'y'=>58.5, 'fontSize'=>12, 'width'=>90, 'appendAsterisk'=>true],
+      ['key'=>'nameInChinese','type'=>'string', 'x'=>65, 'y'=>62, 'fontSize'=>10, 'width'=>100, 'lang'=>'chn'],
+      ['key'=>'hkid','type'=>'string', 'x'=>100, 'y'=>66.9, 'fontSize'=>12, 'width'=>90, 'appendAsterisk'=>true],
+      ['key'=>'ppNum','type'=>'string', 'x'=>65, 'y'=>71.2, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'gender','type'=>'string', 'x'=>100, 'y'=>76.5, 'fontSize'=>12, 'width'=>90, 'appendAsterisk'=>true],
+      ['key'=>'martialStatus','type'=>'string', 'x'=>100, 'y'=>81.7, 'fontSize'=>12, 'width'=>90, 'appendAsterisk'=>true],
+
       // spouse
-      ['key'=>'spouseSurname','type'=>'string', 'x'=>52, 'y'=>100, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'spouseGivenName','type'=>'string', 'x'=>52, 'y'=>104, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'spousePpNum','type'=>'string', 'x'=>52, 'y'=>108, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'spouseName','type'=>'string', 'x'=>90, 'y'=>85.5, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'spouseHkid','type'=>'string', 'x'=>90, 'y'=>89.7, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'spousePpNum','type'=>'string', 'x'=>105, 'y'=>93.9, 'fontSize'=>10, 'width'=>100],
+
       // correspondence
-      ['key'=>'resAddress','type'=>'string', 'x'=>52, 'y'=>112, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'posAddress','type'=>'string', 'x'=>52, 'y'=>116, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'resAddress','type'=>'string', 'x'=>62, 'y'=>98.1, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'posAddress','type'=>'string', 'x'=>90, 'y'=>105.7, 'fontSize'=>10, 'width'=>100],
+
       // position
-      ['key'=>'capacity','type'=>'string', 'x'=>52, 'y'=>120, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'startDateOfEmp','type'=>'string', 'x'=>52, 'y'=>124, 'fontSize'=>10, 'width'=>100, 'fontStyle'=>'B'],
-      ['key'=>'monthlyFixedIncome','type'=>'string', 'x'=>52, 'y'=>128, 'fontSize'=>10, 'width'=>100, 'fontStyle'=>'B'],
-      ['key'=>'monthlyAllowance','type'=>'string', 'x'=>52, 'y'=>132, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'capacity','type'=>'string', 'x'=>70, 'y'=>114.2, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'startDateOfEmp','type'=>'string', 'x'=>10, 'y'=>122.8, 'fontSize'=>10, 'width'=>90, 'appendAsterisk'=>true],
+      ['key'=>'monthlyFixedIncome','type'=>'string', 'x'=>100, 'y'=>131.5, 'fontSize'=>10, 'width'=>90, 'appendAsterisk'=>true],
+
+
+      ['key'=>'monthlyAllowance','type'=>'string', 'x'=>110, 'y'=>135.9, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'fluctuatingIncome','type'=>'string', 'x'=>120, 'y'=>141.1, 'fontSize'=>10, 'width'=>100],
+
+
       // place of residence
-      ['key'=>'placeProvided','type'=>'string', 'x'=>52, 'y'=>136, 'fontSize'=>10, 'width'=>100, 'fontStyle'=>'B'],
-      ['key'=>'addrOfPlace','type'=>'string', 'x'=>52, 'y'=>136, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'natureOfPlace','type'=>'string', 'x'=>52, 'y'=>140, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'rentPaidEr','type'=>'string', 'x'=>52, 'y'=>144, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'rentPaidEe','type'=>'string', 'x'=>52, 'y'=>148, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'rentRefund','type'=>'string', 'x'=>52, 'y'=>152, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'rentPaidErByEe','type'=>'string', 'x'=>52, 'y'=>156, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'overseaIncInd','type'=>'string', 'x'=>52, 'y'=>160, 'fontSize'=>10, 'width'=>100, 'fontStyle'=>'B'],
-      ['key'=>'amtPaidOverseaCo','type'=>'string', 'x'=>52, 'y'=>164, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'nameOfOverseaCo','type'=>'string', 'x'=>52, 'y'=>168, 'fontSize'=>10, 'width'=>100],
-      ['key'=>'addrOfOverseaCo','type'=>'string', 'x'=>52, 'y'=>172, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'placeProvided','type'=>'string', 'x'=>52, 'y'=>148.3, 'fontSize'=>10, 'width'=>100, 'appendAsterisk'=>true],
+      ['key'=>'addrOfPlace','type'=>'string', 'x'=>52, 'y'=>155.5, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'natureOfPlace','type'=>'string', 'x'=>52, 'y'=>160.7, 'fontSize'=>10, 'width'=>100],
+
+      ['key'=>'rentPaidEr','type'=>'string', 'x'=>105, 'y'=>165.9, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'rentPaidEe','type'=>'string', 'x'=>105, 'y'=>171.3, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'rentRefund','type'=>'string', 'x'=>105, 'y'=>176.7, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'rentPaidErByEe','type'=>'string', 'x'=>105, 'y'=>182.1, 'fontSize'=>10, 'width'=>100],
+
+      // Non-hong kong company
+      ['key'=>'overseaIncInd','type'=>'string', 'x'=>52, 'y'=>195, 'fontSize'=>10, 'width'=>100, 'appendAsterisk'=>true],
+      ['key'=>'amtPaidOverseaCo','type'=>'string', 'x'=>57, 'y'=>199.2, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'nameOfOverseaCo','type'=>'string', 'x'=>85, 'y'=>203.4, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'addrOfOverseaCo','type'=>'string', 'x'=>42, 'y'=>207.6, 'fontSize'=>10, 'width'=>100],
+
       // share option
-      ['key'=>'shareBeforeEmp','type'=>'string', 'x'=>52, 'y'=>176, 'fontSize'=>10, 'width'=>100]
+      ['key'=>'shareBeforeEmp','type'=>'string', 'x'=>120.5, 'y'=>223.5, 'fontSize'=>10, 'width'=>100],
+
+      // Footer
+      ['key'=>'signatureName','type'=>'string', 'x'=>92, 'y'=>257.5, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'designation','type'=>'string', 'x'=>100, 'y'=>265.5, 'fontSize'=>10, 'width'=>100],
+      ['key'=>'formDate','type'=>'string', 'x'=>92, 'y'=>273.5, 'fontSize'=>10, 'width'=>100],
+
     ];
-    $fieldList = [];
-    foreach($fields as $field) {
-      $fieldList[] = (object) $field;
-    }
+//
+//
+//    $fieldList = [];
+//
+//
+////    $fs = $irdFormFile->fields;
+////    foreach($fs as $f) {
+////      echo $f->key; nl();
+////    }
+////    dd('ok');
+//
+//
+//    foreach($fields as $field) {
+//      $fieldList[] = (object) $field;
+//
+//      $align = array_key_exists('align', $field) ? $field['align'] : 'L';
+//      $charAlign = array_key_exists( 'charAlign', $field) ? $field['charAlign'] : 'C';
+//      $lang = array_key_exists('lang', $field) ? $field['lang'] : 'eng';
+//      $appendAsterisk = array_key_exists('appendAsterisk', $field) ? $field['appendAsterisk'] : false;
+//
+////      $irdFormFile->fields()->save(new IrdFormFileField([
+////        'key'=>$field['key'],
+////        'type'=>$field['type'],
+////        'x'=>$field['x'],
+////        'y'=>$field['y'],
+////        'font_size'=>$field['fontSize'],
+////        'relative_to'=>'global',
+////        'relative_to_key_id'=>0,
+////        'width'=>$field['width'],
+////        'field_count'=>0,
+////        'align'=>$align,
+////        'char_align'=>$charAlign,
+////        'lang'=>$lang,
+////        'append_asterisk'=>$appendAsterisk,
+////        'remark'=>''
+////      ]));
+//      if($appendAsterisk) {
+//        echo 'key='.$field['key']; nl();
+//        $irdFormFile->fields()->where('key', $field['key'])->update(['append_asterisk'=>true]);
+////        first();
+//
+////        echo $irdFormField->key; nl();
+//      }
+//    }
+//    dd('ok');
+
+    $fieldList = $irdFormFile->fields;
+
     foreach($fieldList as $item) {
       $align = isset($item->align) ? $item->align : 'L';
       $lang = isset($item->lang) ? $item->lang : 'eng';
-      $fontStyle = isset($item->fontStyle) ? $item->fontStyle : '';
+      $fontStyle = isset($item->font_style) ? $item->font_style : '';
       switch ($item->type) {
         case 'string':
+          $text = $data->{$item->key};
+          if(hasChinese($text)) {
+            $lang = 'chn';
+          }
+          $x = $item->x;
+          $y = self::$yOffset + $item->y;
+          $width = $item->width;
+          $fontSize = $item->font_size;
+          $appendAsterisk = isset($item->append_asterisk) ? $item->append_asterisk : false;
+
+          // Is Currency
+          if(!empty($text)) {
+            if($item->to_currency) {
+              $text = toCurrency(str_replace(',', '', $text));
+            }
+          }
+
+          // Append Asterisk
+          if($appendAsterisk) {
+            $x = 100;
+            $width = 90;
+            $align = 'R';
+            $fontSize = 12;
+            $fontStyle = 'B';
+            $text .= ' ****';
+          }
           $pdf->outputText(
-            $item->x,
-            self::$yOffset + $item->y,
-            $item->fontSize,
-            $item->width,
-            $data->{$item->key},
+            $x,
+            $y,
+            $fontSize,
+            $width,
+            $text,
             $align,
             $lang,
+            null,
             $fontStyle
           );
           break;
@@ -98,11 +214,14 @@ class CommencementFormPdfHelper {
 
       }
     }
-
-    if(file_exists($finalFilePath)) {
-      unlink($finalFilePath);
+    if(isset($finalFilePath)) {
+      if (file_exists($finalFilePath)) {
+        unlink($finalFilePath);
+      }
+      $pdf->Output($finalFilePath, 'F');
+    } else {
+      $pdf->Output('commencement_form.pdf');
     }
-    $pdf->Output($finalFilePath, 'F');
     return;
